Cleanup random
